Add recheck buttons after downloading new data from server
Only bug is when you try to download data from server that is already
available, because the download_dataset just returns, but not as a
object which support the .done method and so it triggers an error. Just
need to check the object returned from download_dataset and make sure it
is not 'undefined'.

Maybe implement function to hold the checked_input and return
input_array