Enable looping
When the end of a json is reached it will be restarted (replayed).
This will cause (more) sync problems because the start and end times
are not equal.
Currently the first events of all stations (and coincidences) happen at
the ‘same’ time, they shouldn’t..
Perhaps we need a global start and end time (min/max timestamp).